Refactor: split up templates.py; update logic in rendering prior to fixing recursion issues
