+++ src_py/templatey/_bootstrapping.py
@@ -1,12 +1,28 @@
 from __future__ import annotations
 
+from collections.abc import Mapping
+from collections.abc import Sequence
+from dataclasses import InitVar
 from typing import cast
 
+from templatey._types import TemplateIntersectable
+from templatey._types import TemplateParamsInstance
 from templatey.interpolators import NamedInterpolator
+from templatey.parser import InterpolationConfig
 from templatey.parser import ParsedTemplateResource
 from templatey.templates import TemplateConfig
-from templatey.templates import TemplateIntersectable
 from templatey.templates import template
+
+
+class _VirtualSlotMixin:
+    """We use virtual slots on empty template instances to create
+    provenance / render frame backstops for environment functions
+    that return template instances. This mixin adds a slot to the
+    class that allows it to be stored there.
+    """
+    _TEMPLATEY_VIRTUAL_INSTANCE = True
+    _templatey_virtual_slots: Mapping[str, Sequence[TemplateParamsInstance]]
+    __slots__ = ['_templatey_virtual_slots']
 
 
 @template(
@@ -16,11 +32,24 @@ from templatey.templates import template
         content_verifier=lambda value: True),
     object()
 )
-class EmptyTemplate:
+class EmptyTemplate(_VirtualSlotMixin):
     """This is used as the render stack anchor for values that are
     injected into a function, and therefore have no parent. It is
     special-cased within the render env.
     """
+    virtual_slots: InitVar[Mapping[str, Sequence[TemplateParamsInstance]]]
+
+    def __post_init__(
+            self,
+            virtual_slots: Mapping[str, Sequence[TemplateParamsInstance]]):
+        self._templatey_virtual_slots = virtual_slots
+
+    def __getattr__(self, key: str):
+        result = self._templatey_virtual_slots.get(key)
+        if result is None:
+            raise AttributeError(key)
+        else:
+            return result
 
 
 PARSED_EMPTY_TEMPLATE = ParsedTemplateResource(
@@ -32,4 +61,5 @@ PARSED_EMPTY_TEMPLATE = ParsedTemplateResource(
     function_names=frozenset(),
     function_calls={})
 EMPTY_TEMPLATE_XABLE = cast(type[TemplateIntersectable], EmptyTemplate)
-EMPTY_TEMPLATE_INSTANCE = EmptyTemplate()
+EMPTY_TEMPLATE_INSTANCE = EmptyTemplate(virtual_slots={})
+EMPTY_INTERPOLATION_CONFIG = InterpolationConfig()
