+++ src_py/templatey/templates.py
@@ -2,91 +2,47 @@ from __future__ import annotations
 
 import functools
 import inspect
-import itertools
 import logging
-import operator
 import sys
 import typing
 from collections import ChainMap
-from collections import defaultdict
-from collections import deque
 from collections import namedtuple
 from collections.abc import Callable
 from collections.abc import Collection
 from collections.abc import Iterable
-from collections.abc import Iterator
 from collections.abc import Mapping
-from collections.abc import MutableMapping
 from collections.abc import Sequence
-from contextlib import contextmanager
-from contextvars import ContextVar
-from copy import copy
 from dataclasses import _MISSING_TYPE
-from dataclasses import KW_ONLY
 from dataclasses import Field
-from dataclasses import InitVar
 from dataclasses import dataclass
 from dataclasses import field
 from dataclasses import fields
-from itertools import tee as tee_iterable
-from random import Random
 from textwrap import dedent
 from types import EllipsisType
 from types import FrameType
-from types import UnionType
 from typing import Annotated
 from typing import Any
-from typing import ClassVar
 from typing import Literal
-from typing import NamedTuple
 from typing import Protocol
-from typing import Self
-from typing import cast
 from typing import dataclass_transform
 from typing import overload
-from weakref import ref
 
 from docnote import ClcNote
 from typing_extensions import TypeIs
 
 from templatey._annotations import InterfaceAnnotation
 from templatey._annotations import InterfaceAnnotationFlavor
+from templatey._forwardrefs import ForwardRefLookupKey
+from templatey._forwardrefs import _ForwardRefGeneratingNamespaceLookup
+from templatey._forwardrefs import extract_frame_scope_id
+from templatey._forwardrefs import resolve_forward_references
+from templatey._signature import TemplateSignature
+from templatey._types import TemplateIntersectable
+from templatey._types import TemplateParamsInstance
 from templatey.interpolators import NamedInterpolator
-from templatey.parser import InterpolatedFunctionCall
 from templatey.parser import InterpolationConfig
-from templatey.parser import NestedContentReference
-from templatey.parser import NestedVariableReference
-from templatey.parser import ParsedTemplateResource
 
-if typing.TYPE_CHECKING:
-    from _typeshed import DataclassInstance
-else:
-    DataclassInstance = object
-
-# Technically this should be an intersection type with both the
-# _TemplateIntersectable from templates and the DataclassInstance returned by
-# the dataclass transform. Unfortunately, type intersections don't yet exist in
-# python, so we have to resort to this (overly broad) type
-TemplateParamsInstance = DataclassInstance
-type TemplateClass = type[TemplateParamsInstance]
 logger = logging.getLogger(__name__)
-# Defining this as a contextvar basically just for testing purposes. We want
-# everything else to get the default global.
-_PENDING_FORWARD_REFS: ContextVar[
-    dict[ForwardRefLookupKey, set[TemplateClass]]] = ContextVar(
-        '_PENDING_FORWARD_REFS', default=defaultdict(set))  # noqa: B039
-
-# Note: we don't need cryptographically secure IDs here, so let's preserve
-# entropy (might also be faster, dunno). Also note: the only reason we're
-# using a contextvar here is so that we can theoretically replace it with
-# a deterministic seed during testing (if we run into flakiness due to
-# non-determinism)
-_ID_PRNG: ContextVar[Random] = ContextVar('_ID_PRNG', default=Random())  # noqa: B039, S311
-_ID_BITS = 128
-# This is used by ``anchor_closure_scope`` to assign a scope ID to templates.
-# It's used for all templates, but will only be non-null inside a closure.
-_CURRENT_SCOPE_ID: ContextVar[int | None] = ContextVar(
-    '_CURRENT_SCOPE_ID', default=None)
 
 
 # Technically, these should use the TemplateIntersectable from templates.py,
@@ -100,30 +56,6 @@ type Var[T] = Annotated[
 type Content[T] = Annotated[
     T,
     InterfaceAnnotation(InterfaceAnnotationFlavor.CONTENT)]
-
-
-type _SlotAnnotation = (
-    TemplateClass
-    | UnionType
-    | type[ForwardReferenceProxyClass])
-
-
-class TemplateIntersectable(Protocol):
-    """This is the actual template protocol, which we would
-    like to intersect with the TemplateParamsInstance, but cannot.
-    Primarily included for documentation.
-    """
-    _templatey_config: ClassVar[TemplateConfig]
-    # Note: whatever kind of object this is, it needs to be understood by the
-    # template loader defined in the template environment. It would be nice for
-    # this to be a typvar, but python doesn't currently support typevars in
-    # classvars
-    _templatey_resource_locator: ClassVar[object]
-    _templatey_signature: ClassVar[TemplateSignature]
-    # Oldschool here for performance reasons; otherwise this would be a dict.
-    # Field names match the field names from the params; the value is gathered
-    # from the metadata value on the field.
-    _templatey_prerenderers: ClassVar[NamedTuple]
 
 
 def is_template_class(cls: type) -> TypeIs[type[TemplateIntersectable]]:
@@ -383,1051 +315,6 @@ class TemplateConfig[T: type, L: object]:
             ''')]
 
 
-@dataclass(slots=True, frozen=True)
-class TemplateProvenanceNode:
-    """TemplateProvenanceNode instances are unique to the exact location
-    on the exact render tree of a particular template instance. If the
-    template instance gets reused within the same render tree, it will
-    have multiple provenance nodes. And each different slot in an
-    enclosing template will have a separate provenance node, potentially
-    with different namespace overrides.
-
-    This is used both during function execution (to calculate the
-    concrete value of any parameters), and also while flattening the
-    render tree.
-
-    Also note that the root template, **along with any templates
-    injected into the render by environment functions,** will have an
-    empty list of provenance nodes.
-
-    Note that, since overrides from the enclosing template can come
-    exclusively from the template body -- and are therefore shared
-    across all nested children of the same slot -- they don't get stored
-    within the provenance, since we'd require access to the template
-    bodies, which we don't yet have.
-    """
-    encloser_slot_key: str
-    encloser_slot_index: int
-    # The reason to have both the instance and the instance ID is so that we
-    # can have hashability of the ID while not imposing an API on the instances
-    instance_id: TemplateInstanceID
-    instance: TemplateParamsInstance = field(compare=False)
-
-
-class TemplateProvenance(tuple[TemplateProvenanceNode]):
-
-    def bind_content(
-            self,
-            name: str,
-            template_preload: dict[TemplateClass, ParsedTemplateResource]
-            ) -> object:
-        """Use this to calculate a concrete value for use in rendering.
-        This walks up the provenance stack, recursively looking for any
-        overrides to the content. If none are found, it returns the
-        value from the childmost instance in the provenance.
-        """
-        # We use the literal ellipsis type as a sentinel for values not being
-        # added yet, so we might as well just continue the trend!
-        current_provenance_node = self[-1]
-        value = getattr(current_provenance_node.instance, name, ...)
-        encloser_param_name = name
-        encloser_slot_key = current_provenance_node.encloser_slot_key
-        for encloser in reversed(self[0:-1]):
-            template_class = type(encloser.instance)
-            encloser_template = template_preload[template_class]
-            encloser_overrides = (
-                encloser_template.slots[encloser_slot_key].params)
-
-            if encloser_param_name in encloser_overrides:
-                value = encloser_overrides[encloser_param_name]
-
-                if isinstance(value, NestedContentReference):
-                    encloser_slot_key = encloser.encloser_slot_key
-                    encloser_param_name = value.name
-                    value = ...
-                else:
-                    break
-            else:
-                break
-
-        if value is ...:
-            raise KeyError(
-                'No value found for content with name at slot!',
-                self[-1].instance, name)
-
-        return value
-
-    def bind_variable(
-            self,
-            name: str,
-            template_preload: dict[TemplateClass, ParsedTemplateResource]
-            ) -> object:
-        """Use this to calculate a concrete value for use in rendering.
-        This walks up the provenance stack, recursively looking for any
-        overrides to the variable. If none are found, it returns the
-        value from the childmost instance in the provenance.
-        """
-        # We use the literal ellipsis type as a sentinel for values not being
-        # added yet, so we might as well just continue the trend!
-        current_provenance_node = self[-1]
-        value = getattr(current_provenance_node.instance, name, ...)
-        encloser_param_name = name
-        encloser_slot_key = current_provenance_node.encloser_slot_key
-        for encloser in reversed(self[0:-1]):
-            template_class = type(encloser.instance)
-            encloser_template = template_preload[template_class]
-            encloser_overrides = (
-                encloser_template.slots[encloser_slot_key].params)
-
-            if encloser_param_name in encloser_overrides:
-                value = encloser_overrides[encloser_param_name]
-
-                if isinstance(value, NestedVariableReference):
-                    encloser_slot_key = encloser.encloser_slot_key
-                    encloser_param_name = value.name
-                    value = ...
-                else:
-                    break
-            else:
-                break
-
-        if value is ...:
-            raise KeyError(
-                'No value found for variable with name at slot!',
-                self[-1].instance, name)
-
-        return value
-
-
-@contextmanager
-def anchor_closure_scope():
-    """We strongly recommend against defining templates within a
-    closure, as it can cause a number of fragility issues, and just
-    generally makes less sense than defining templates at the module
-    level. However, if you absolutely must create a new template within
-    a closure, you must use ``anchor_closure_scope`` to give the
-    templates a known closure scope. Can be used either as a decorator
-    or a context:
-
-    > Decorator usage
-    __embed__: 'code/python'
-        @anchor_closure_scope()
-        def my_func():
-            # template definition goes here
-            ...
-
-    > Context manager usage
-    __embed__: 'code/python'
-        def my_other_func():
-            with anchor_closure_scope():
-                # template definition goes here
-                ...
-    """
-    token = _CURRENT_SCOPE_ID.set(_create_templatey_id())
-    try:
-        yield
-    finally:
-        _CURRENT_SCOPE_ID.reset(token)
-
-
-def _create_templatey_id() -> int:
-    """Templatey IDs are unique identifiers (theoretically, absent
-    birthday collisions) that we currently use in two places:
-    ++  as a scope ID, which is used when defining templates in closures
-    ++  for giving slot tree nodes a unique reference target for
-        recursion loops while copying and merging, which is more robust
-        than ``id(target)`` and can be transferred via dataclass field
-        into cloned/copied/merged nodes.
-    """
-    prng = _ID_PRNG.get()
-    return prng.getrandbits(_ID_BITS)
-
-
-@dataclass(frozen=True, slots=True)
-class ForwardRefLookupKey:
-    """We use this to find all possible ForwardRefLookupKey instances
-    for a particular pending template.
-
-    Note that, by definition, forward references can only happen in two
-    situations:
-    ++  within the same module or closure, by something declared later
-        on during execution
-    ++  because of something hidden behind a ``if typing.TYPE_CHECKING``
-        block in imports
-    (Any other scenario would result in import failures preventing the
-    module's execution).
-
-    Names imported behind ``TYPE_CHECKING`` blocks can **only** be
-    resolved using explicit helpers in the ``@template`` decorator.
-    (TODO: need to add those!). There's just no way around that one;
-    by definition, it's a circular import, and the name isn't available
-    at runtime. So you need an escape hatch.
-
-    Therefore, unless passed in an explicit module name because of the
-    aforementioned escape hatch, these must always happen from within
-    the same module as the template itself.
-
-    Furthermore, we make one assumption here for the purposes of
-    doing as much work as possible at import time, ahead of the first
-    call to render a template: that the enclosing template references
-    the nested template by the nested template's proper name, and
-    doesn't rename it.
-
-    The only workaround for a renamed nested template would be to
-    create a dedicated resolution function, to be called at first
-    render time, that re-inspects the template's type annotations, and
-    figures out exactly what type it uses at that point in time. That's
-    a mess, so.... we'll punt on it.
-    """
-    module: str
-    name: str
-    scope_id: int | None
-
-
-# Note: the ordering here is to emphasize the fact that the slot
-# name is on the ENCLOSING template, but the slot type is from the
-# NESTED template
-class _SlotTreeRoute[T: _SlotTreeNode](tuple[str, TemplateClass, T]):
-    """An individual route on the slot tree is defined by the attribute
-    name for the slot, the slot type, and the subtree from the slot
-    class.
-
-    These are optimized for the non-union case. Traversing the slot tree
-    with union types will result in a bunch of unnecessary comparisons
-    against slot names of different slot types.
-
-    Note that slot tree routes always have a concrete slot name and slot
-    type, regardless of whether they're in a pending or concrete tree.
-    The reason is simple: in a pending tree, all of the pending classes
-    are dead-end nodes, and define their insertion points using just
-    the string of the slot name, and nothing else.
-    """
-    @classmethod
-    def new(
-            cls,
-            slot_name: str,
-            slot_type: TemplateClass,
-            subtree: T,
-            ) -> _SlotTreeRoute[T]:
-        return cls((slot_name, slot_type, subtree))
-
-    @property
-    def subtree(self) -> T:
-        """This is slower than directly accessing the tuple values, but
-        it makes for clearer code during tree building, where
-        performance isn't quite so critical.
-        """
-        return self[2]
-
-    @property
-    def slot_path(self) -> tuple[str, TemplateClass]:
-        return self[0:2]
-
-
-@dataclass(slots=True)
-class _SlotTreeNode[T: _SlotTreeNode](list[_SlotTreeRoute[T]]):
-    """The purpose of the slot tree is to precalculate what sequences of
-    getattr() calls we need to traverse to arrive at every instance of a
-    particular slot type for a given template, including all nested
-    templates.
-
-    **These are optimized for rendering, not for template declaration.**
-    Also note that these are optimized for slots that are not declared
-    as type unions; type unions will result in a number of unnecessary
-    comparisons against the routes of the other slot types in the union.
-
-    The reason this is useful is for batching during rendering. This is
-    important for function calls: it allows us to pre-execute all env
-    func calls for a template before we start rendering it. In the
-    future, it will also serve the same role for discovering the actual
-    template types for dynamic slots, allowing us to load the needed
-    template types in advance.
-
-    An individual node on the slot tree is a list of all possible
-    attribute names (as ``_SlotTreeRoute``s) that a particular search
-    pass needs to check for a given instance. Note that **all** of the
-    attributes must be searched -- hence using an iteration-optimized
-    list instead of a mapping.
-    """
-    routes: InitVar[Iterable[_SlotTreeRoute[T]] | None] = None
-
-    _: KW_ONLY
-    # We use this to limit the number of entries we need in the transmogrifier
-    # lookup during tree merging/copying
-    is_recursive: bool = False
-    # We use this to differentiate between dissimilar unions, one which
-    # continues on to a subtree, and one which ends here with the target
-    # instance
-    is_terminus: bool = False
-
-    id_: int = field(default_factory=_create_templatey_id)
-
-    # We use this to make the logic cleaner when merging trees, but we want
-    # the faster performance of the tuple when actually traversing the tree
-    _index_by_slot_path: dict[tuple[str, TemplateClass], int] = field(
-            init=False, repr=False)
-
-    def __post_init__(
-            self,
-            routes: Iterable[_SlotTreeRoute[T]] | None):
-        # Explicit instead of super because... idunno, we're breaking things
-        # somehow
-        if routes is None:
-            list.__init__(self)
-        else:
-            list.__init__(self, routes)
-
-        self._index_by_slot_path = {
-            route.slot_path: index for index, route in enumerate(self)}
-
-    def empty_clone(
-            self,
-            *,
-            fields_to_skip: set[str] | frozenset[str] = frozenset()
-            ) -> Self:
-        """This creates a clone of the node without any routes. Useful
-        for merging and copying, where you need to do some manual
-        transform of the content.
-
-        Note that this is almost the same as dataclasses.replace, with
-        the exception that we create shallow copies of attributes
-        instead of preserving them.
-        """
-        kwargs = {}
-        for dc_field in fields(self):
-            if dc_field.init and dc_field.name not in fields_to_skip:
-                # Note: the copy here is important for any mutable values,
-                # notably the insertion_slot_names.
-                kwargs[dc_field.name] = copy(getattr(self, dc_field.name))
-
-        return type(self)(**kwargs)
-
-    def merge_fields_only(
-            self,
-            other: _SlotTreeNode,
-            *,
-            fields_to_skip: set[str] | frozenset[str] = frozenset()):
-        """Updates the current node, merging in all non-init field
-        values from other, using |=. Only merges values that exist on
-        the current node, allowing for transformation between pending
-        and concrete node types.
-
-        Leaves the ID of the current node unchanged.
-        """
-        # Always skip ``id_`` -- always preserve the original ID!
-        fields_to_skip = fields_to_skip | {'id_'}
-        missing = object()
-
-        for dc_field in fields(self):
-            if dc_field.init and dc_field.name not in fields_to_skip:
-                current_value = getattr(self, dc_field.name)
-                other_value = getattr(other, dc_field.name, missing)
-
-                if other_value is not missing:
-                    setattr(
-                        self,
-                        dc_field.name,
-                        operator.ior(current_value, other_value))
-
-    @property
-    def requires_transmogrification(self) -> bool:
-        """This determines whether or not copies of the tree require
-        some post-processing to make sure that the tree STRUCTURE is
-        the same. It is used by both copying and merging trees to make
-        sure that the transmogrification lookup is as sparse as
-        possible.
-
-        For the base class, we simply wrap ``is_recursive``, but for
-        the pending tree derived class, we also check other stuff --
-        hence the wrapping.
-        """
-        return self.is_recursive
-
-    def append(self, route: _SlotTreeRoute[T]):
-        """This has slightly different semantics to normal list
-        appending if you're trying to append a duplicate slot path:
-        ++  if it's identical to the existing path (ie, it targets the
-            same node), we do nothing
-        ++  if it's not identical, we error
-        """
-        slot_path = (route[0], route[1])
-        slot_index = self._index_by_slot_path.get(slot_path)
-
-        if slot_index is not None:
-            if self[slot_index].subtree is route.subtree:
-                return
-            else:
-                raise ValueError(
-                    'Templatey internal error: attempt to append duplicate '
-                    + 'slot name for same slot type, but a different subtree! '
-                    + 'Please search for / report issue to github along with '
-                    + 'a traceback.')
-
-        list.append(self, route)
-        self._index_by_slot_path[slot_path] = len(self) - 1
-
-    def has_route_for(
-            self,
-            slot_name: str,
-            slot_type: TemplateClass
-            ) -> bool:
-        return (slot_name, slot_type) in self._index_by_slot_path
-
-    def get_route_for(
-            self,
-            slot_name: str,
-            slot_type: TemplateClass
-            ) -> _SlotTreeRoute[T]:
-        return self[self._index_by_slot_path[(slot_name, slot_type)]]
-
-    def rewrite_route_for(
-            self,
-            slot_name: str,
-            slot_type: TemplateClass,
-            new_route: _SlotTreeRoute[T]
-            ) -> None:
-        dest_index = self._index_by_slot_path[(slot_name, slot_type)]
-        list.__setitem__(self, dest_index, new_route)
-
-    def stringify(
-            self,
-            *,
-            depth=0,
-            _encountered_ids: frozenset[int] = frozenset()
-            ) -> str:
-        """Creates a pretty-print-style string representation of the
-        node and all its nested nodes, recursively.
-        """
-        indentation = '    ' * depth
-
-        to_join = []
-        for dc_field in fields(self):
-            if dc_field.init:
-                field_name = dc_field.name
-                to_join.append(
-                    f'{indentation}{field_name}: {getattr(self, field_name)}')
-
-        if self.id_ in _encountered_ids:
-            to_join.append(
-                f'{indentation}... Recursion detected; omitting subtrees.')
-
-        else:
-            for route in self:
-                to_join.append(
-                    f'{indentation}++  {route[0:2]}')
-                to_join.append(route[2].stringify(
-                    depth=depth + 1,
-                    _encountered_ids=_encountered_ids | {self.id_}))
-
-        return '\n'.join(to_join)
-
-    def is_equivalent(
-            self,
-            other: _SlotTreeNode,
-            *,
-            _previous_encounters: dict[int, int] | None = None
-            ) -> bool:
-        """This compares two slot tree nodes recursively, ignoring IDs.
-        If they are otherwise identical -- including the structure of
-        any recursive loops -- returns True.
-
-        Note that this is optimized for maintainability and not
-        performance. Its primary intended use is in testing.
-        """
-        previous_encounters: dict[int, int]
-        if _previous_encounters is None:
-            previous_encounters = {}
-        else:
-            # Note that the copy here is important because otherwise we'd be
-            # mutating state even in other tree branches which might not have
-            # encountered us (since we reuse this in the recursive case)
-            previous_encounters = {**_previous_encounters}
-
-        # These two cases are trivial: first, if the types aren't the same,
-        # they're not equivalent, period. Second, if we already encountered
-        # that node's ID, that means we're in a recursive loop, and the two
-        # IDs must match.
-        if type(self) is not type(other):
-            return False
-        if other.id_ in previous_encounters:
-            return previous_encounters[other.id_] == self.id_
-
-        # First check all the fields, since this should in theory be quick
-        fields_match: bool = True
-        for dc_field in fields(self):
-            if dc_field.init and dc_field.name != 'id_':
-                fields_match &= (
-                    getattr(self, dc_field.name)
-                    == getattr(other, dc_field.name))
-        if not fields_match:
-            return False
-
-        # Now make sure the routes are the same. This lets us simplify the
-        # recursive comparison logic; we don't need any checks to make sure
-        # that there weren't any leftover routes on either self or other.
-        # It also lets us short-circuit without recursion if they don't match,
-        # but that's not the primary motivation behind it.
-        self_routes = set(self._index_by_slot_path)
-        other_routes = set(other._index_by_slot_path)
-        if self_routes != other_routes:
-            return False
-
-        # Now finally we're getting to recursive territory.
-        previous_encounters[other.id_] = self.id_
-        for slot_path in self._index_by_slot_path:
-            self_index = self._index_by_slot_path[slot_path]
-            self_subtree = self[self_index].subtree
-            other_index = self._index_by_slot_path[slot_path]
-            other_subtree = other[other_index].subtree
-
-            # Recursion is glorious as long as you don't need to worry about
-            # performance! Too bad we can't use it during rendering (because
-            # it's way too slow)
-            if not self_subtree.is_equivalent(
-                other_subtree,
-                _previous_encounters=previous_encounters
-            ):
-                return False
-
-        return True
-
-    def extend(self, routes: Iterable[_SlotTreeRoute[T]]):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def __delitem__(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def __setitem__(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def clear(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def copy(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def insert(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def pop(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def remove(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def reverse(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def __imul__(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-    def __iadd__(self, *args, **kwargs):
-        raise ZeroDivisionError('Templatey internal error: not implemented!')
-
-
-type _ConcreteSlotTreeNode = _SlotTreeNode[_ConcreteSlotTreeNode]
-
-
-@dataclass(kw_only=True, slots=True)
-class _PendingSlotTreeNode(_SlotTreeNode['_PendingSlotTreeNode']):
-    # Note: the str is the slot_name that the route needs to be inserted under
-    insertion_slot_names: set[str] = field(default_factory=set)
-
-    @property
-    def is_insertion_point(self) -> bool:
-        return bool(self.insertion_slot_names)
-
-    @property
-    def requires_transmogrification(self) -> bool:
-        return self.is_recursive
-
-
-type TemplateInstanceID = int
-type GroupedTemplateInvocations = dict[TemplateClass, list[TemplateProvenance]]
-type TemplateLookupByID = dict[TemplateInstanceID, TemplateParamsInstance]
-# Note that there's no need for an abstract version of this, at least not right
-# now, because in order to calculate it, we'd need to know the template body,
-# which doesn't happen until we already know the template instance, which means
-# we can skip ahead to the concrete version.
-type EnvFuncInvocation = tuple[TemplateProvenance, InterpolatedFunctionCall]
-
-
-@dataclass(slots=True)
-class TemplateSignature:
-    """This class stores the processed interface based on the params.
-    It gets used to compare with the TemplateParse to make sure that
-    the two can be used together.
-
-    Not meant to be created directly; instead, you should use the
-    TemplateSignature.new() convenience method.
-    """
-    # It's nice to have this available, especially when resolving forward refs,
-    # but unlike eg the slot tree, it's trivially easy for us to avoid GC
-    # loops within the signature
-    template_cls_ref: ref[TemplateClass]
-    _forward_ref_lookup_key: ForwardRefLookupKey
-
-    slot_names: frozenset[str]
-    var_names: frozenset[str]
-    content_names: frozenset[str]
-
-    # Note that these contain all included types, not just the ones on the
-    # outermost layer that are associated with the signature. In other words,
-    # they include the flattened recursion of all included slots, all the way
-    # down the tree
-    _slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode]
-    _pending_ref_lookup: dict[ForwardRefLookupKey, _PendingSlotTreeContainer]
-
-    # I really don't like that we need to remember to recalculate this every
-    # time we update the slot tree lookup, but for rendering performance
-    # reasons we want this to be precalculated before every call to render.
-    included_template_classes: frozenset[TemplateClass] = field(init=False)
-
-    def __post_init__(self):
-        self.refresh_included_template_classes_snapshot()
-        self.refresh_pending_forward_ref_registration()
-
-    def refresh_included_template_classes_snapshot(self):
-        """Call this when resolving forward references to apply any
-        changes made to the slot tree to the template classes snapshot
-        we use for increased render performance.
-        """
-        template_cls = self.template_cls_ref()
-        if template_cls is None:
-            raise RuntimeError(
-                'Template class was garbage collected before template '
-                + 'signature, and then signature asked to refresh included '
-                + 'classes snapshot?!')
-
-        self.included_template_classes = frozenset(
-            {template_cls, *self._slot_tree_lookup})
-
-    def refresh_pending_forward_ref_registration(self):
-        """Call this after having resolved forward references (or when
-        initially constructing the template signature) to register the
-        template class as requiring its forward refs. This is what
-        plumbs up the notification code to actually initiate resolving.
-        """
-        template_cls = self.template_cls_ref()
-        if template_cls is None:
-            raise RuntimeError(
-                'Template class was garbage collected before template '
-                + 'signature, and then signature asked to refresh pending '
-                + 'forward ref registration?!')
-
-        forward_ref_registry = _PENDING_FORWARD_REFS.get()
-        for forward_ref in self._pending_ref_lookup:
-            forward_ref_registry[forward_ref].add(template_cls)
-
-    @classmethod
-    def new(
-            cls,
-            template_cls: type,
-            slots: dict[str, _SlotAnnotation],
-            vars_: dict[str, type | type[ForwardReferenceProxyClass]],
-            content: dict[str, type | type[ForwardReferenceProxyClass]],
-            *,
-            forward_ref_lookup_key: ForwardRefLookupKey
-            ) -> TemplateSignature:
-        """Create a new TemplateSignature based on the gathered slots,
-        vars, and content. This does all of the convenience calculations
-        needed to populate the semi-redundant fields.
-        """
-        slot_names = frozenset(slots)
-        var_names = frozenset(vars_)
-        content_names = frozenset(content)
-
-        # Quick refresher: our goal here is to construct a lookup that gets
-        # us a route to every instance of a particular template type. In other
-        # words, we want to be able to check a template type, and then see all
-        # possible getattr() sequences that arrive at an instance of that
-        # template type.
-        tree_wip: dict[TemplateClass, _ConcreteSlotTreeNode]
-        tree_wip = defaultdict(_SlotTreeNode)
-        pending_ref_lookup: \
-            dict[ForwardRefLookupKey, _PendingSlotTreeContainer] = {}
-
-        concrete_slot_defs, pending_ref_defs = cls._normalize_slot_defs(
-            slots.items())
-
-        # Note that order between concrete_slot_defs and pending_ref_defs
-        # DOES matter here. Concrete has to come first, because we need to
-        # discover all reference loops back to the template class before
-        # we can fully define the pending trees.
-        for slot_name, slot_type in concrete_slot_defs:
-            # In the simple recursion case -- a template defines a slot of its
-            # own class -- we can immediately create a reference loop without
-            # any pomp nor circumstance.
-            # Also: yes, this is resolved at annotation time, and not a forward
-            # ref!
-            if slot_type is template_cls:
-                recursive_slot_tree = tree_wip[slot_type]
-                recursive_slot_tree.is_terminus = True
-                recursive_slot_tree.is_recursive = True
-                recursive_slot_route = _SlotTreeRoute.new(
-                    slot_name,
-                    slot_type,
-                    recursive_slot_tree)
-                recursive_slot_tree.append(recursive_slot_route)
-
-            # Remember that we expanded the union already, so this is
-            # guaranteed to be a single concrete ``slot_type``.
-            else:
-                offset_tree = _PendingSlotTreeNode(
-                    insertion_slot_names={slot_name})
-                _update_encloser_with_trees_from_slot(
-                    template_cls,
-                    tree_wip,
-                    pending_ref_lookup,
-                    forward_ref_lookup_key,
-                    slot_type,
-                    offset_tree,)
-
-        # Note that by "direct" we mean, immediate nested children of the
-        # current template_cls, for which we're constructing a signature,
-        # and NOT any nested children of nested concrete slots.
-        # These plain pending refs are very straightforward, since we don't
-        # know anything about them yet (by definition; they're forward refs!).
-        for (
-            direct_pending_slot_name,
-            direct_forward_ref_lookup_key
-        ) in pending_ref_defs:
-            existing_pending_tree = pending_ref_lookup.get(
-                direct_forward_ref_lookup_key)
-
-            if existing_pending_tree is None:
-                dest_insertion = _PendingSlotTreeNode(
-                    insertion_slot_names={direct_pending_slot_name})
-                pending_ref_lookup[direct_forward_ref_lookup_key] = (
-                    _PendingSlotTreeContainer(
-                        pending_slot_type=direct_forward_ref_lookup_key,
-                        pending_root_node=dest_insertion))
-                # Note that we need to include any recursion loops that end
-                # up back at the template class, since they would ALSO have
-                # the same insertion points. Helpfully, we can just merge in
-                # any existing tree for that.
-                existing_recursive_self_tree = tree_wip.get(
-                    template_cls,
-                    _SlotTreeNode())
-                _merge_into_slot_tree(
-                    template_cls,
-                    None,
-                    dest_insertion,
-                    existing_recursive_self_tree)
-
-            else:
-                (existing_pending_tree
-                    .pending_root_node
-                    .insertion_slot_names.add(direct_pending_slot_name))
-
-        # Oh thank god.
-        tree_wip.default_factory = None
-        return cls(
-            _forward_ref_lookup_key=forward_ref_lookup_key,
-            template_cls_ref=ref(template_cls),
-            slot_names=slot_names,
-            var_names=var_names,
-            content_names=content_names,
-            _slot_tree_lookup=tree_wip,
-            _pending_ref_lookup=pending_ref_lookup)
-
-    @classmethod
-    def _normalize_slot_defs(
-            cls,
-            slot_defs: Iterable[tuple[str, _SlotAnnotation]]
-            ) -> tuple[
-                list[tuple[str, TemplateClass]],
-                list[tuple[str, ForwardRefLookupKey]]]:
-        """The annotations we get "straight off the tap" (so to speak)
-        of the template class can be:
-        ++  unions
-        ++  type aliases, though we don't quite support these yet
-            (though this function should make it relatively
-            straightforward to do so)
-        ++  concrete backrefs to existing templates
-        ++  pending forward refs to not-yet-defined templates
-
-        This function is responsible for giving us a clean, uniform way
-        of representing them. First, it expands all unions etc into
-        multiple slot paths. Second, it splits the concrete from the
-        pending nodes, returning them separately.
-        """
-        concrete_slots = []
-        pending_refs = []
-
-        flattened_defs: \
-            list[tuple[
-                str,
-                TemplateClass | type[ForwardReferenceProxyClass]]] = []
-        for slot_name, slot_annotation in slot_defs:
-            # Note that this still might contain a heterogeneous mix of
-            # template classes and forward refs! Hence flattening first.
-            if isinstance(slot_annotation, UnionType):
-                for slot_type in slot_annotation.__args__:
-                    flattened_defs.append((slot_name, slot_type))
-
-            else:
-                flattened_defs.append((slot_name, slot_annotation))
-
-        # Again, this looks redundant at first glance, but the point was to
-        # normalize unions into single types, whether concrete or pending
-        for slot_name, flattened_slot_annotation in flattened_defs:
-            if is_forward_reference_proxy(flattened_slot_annotation):
-                forward_ref = flattened_slot_annotation.REFERENCE_TARGET
-                pending_refs.append((slot_name, forward_ref))
-
-            else:
-                concrete_slots.append((slot_name, flattened_slot_annotation))
-
-        return concrete_slots, pending_refs
-
-    def extract_function_invocations(
-            self,
-            root_template_instance: TemplateParamsInstance,
-            template_preload: dict[TemplateClass, ParsedTemplateResource]
-            ) -> list[EnvFuncInvocation]:
-        """Looks at all included abstract function invocations, and
-        generates lists of their concrete invocations, based on both the
-        actual values of slots at the template instances, as well as the
-        template definition provided in template_preload.
-        """
-        invocations: list[EnvFuncInvocation] = []
-
-        # Things to keep in mind when reading the following code:
-        # ++  it may be easiest to step through using an example, perhaps from
-        #     the test suite.
-        # ++  enclosing template classes can have multiple slots with the same
-        #     nested template class. We call these "parallel slots" in this
-        #     function; they're places where the slot search tree needs to
-        #     split into multiple branches
-        # ++  the combinatorics here can be confusing, because we can have both
-        #     multiple instances in each slot AND multiple slots for each
-        #     template class
-        # ++  multiple instances per slot branch the instance search tree, but
-        #     multiple slots per template class branch the slot search tree.
-        #     However, we need to exhaust both search trees when finding all
-        #     relevant provenances. Therefore, we have to periodically refresh
-        #     one of the search trees, whenever we step to a sibling on the
-        #     other tree
-
-        # The goal of the parallel slot backlog is to keep track of which other
-        # SLOTS (attributes! not instances!) at a particular instance are ALSO
-        # on the search path, and therefore need to be returned to after the
-        # current branch has been exhausted.
-        # The parallel slot backlog is a stack of stacks. The outer stack
-        # corresponds to the depth in the slot tree. The inner stack
-        # contains all remaining slots to search for a particular depth.
-        parallel_slot_backlog_stack: list[list[_SlotTreeRoute]]
-
-        # The goal of the instance backlog stack is to keep track of all the
-        # INSTANCES (not slots / attributes!) that are also on the search path,
-        # and therefore need to be returned to after the current branch has
-        # been exhausted. Its deepest level gets refreshed from the instance
-        # history stack every time we move on to a new parallel slot.
-        # Similar (but different) to the parallel slot backlog, the instance
-        # backlog is a stack of **queues** (it's important to preserve order,
-        # since slot members are by definition ordered). The outer stack
-        # again corresponds to the depth in the slot tree, and the inner queue
-        # contains all of the remaining instances to search for a particular
-        # depth.
-        instance_backlog_stack: list[deque[TemplateProvenanceNode]]
-
-        # The instance history stack is similar to the backlog stack; however,
-        # it is not mutated on a particular level. Use it to "refresh" the
-        # instance backlog stack for parallel slots.
-        instance_history_stack: list[tuple[TemplateProvenanceNode, ...]]
-
-        # These are all used per-iteration, and don't keep state across
-        # iterations.
-        nested_slot_name: str
-        nested_slot_routes: _ConcreteSlotTreeNode
-        nested_instances: Sequence[TemplateParamsInstance]
-
-        # This is used in multiple loop iterations, plus at the end to add any
-        # function calls for the root instance.
-        root_provenance_node = TemplateProvenanceNode(
-            encloser_slot_key='',
-            encloser_slot_index=-1,
-            instance_id=id(root_template_instance),
-            instance=root_template_instance)
-
-        # Keep in mind that the slot tree contains all included slot classes
-        # (recursively), not just the ones at the root_template_instance.
-        # Our goal here is:
-        # 1.. find all template classes with abstract function calls
-        # 2.. build provenances for all invocations of those template classes
-        # 3.. combine (product) those provenances with all of the abstract
-        #     function calls at that template class
-        for template_class, root_nodes in self._slot_tree_lookup.items():
-            abstract_calls = template_preload[template_class].function_calls
-
-            # Constructing a provenance is relatively expensive, so we only
-            # want to do it if we actually have some function calls within the
-            # template
-            if abstract_calls:
-                provenances: list[TemplateProvenance] = []
-                parallel_slot_backlog_stack = [list(root_nodes)]
-                instance_history_stack = [(root_provenance_node,)]
-                instance_backlog_stack = [deque(instance_history_stack[0])]
-
-                # Our overall strategy here is to let the instance stack be
-                # the primary driver. Only mutate other stacks when we're done
-                # with a particular level in the instance backlog!
-                while instance_backlog_stack:
-                    # If there's nothing left on the current level of the
-                    # instance backlog, there are a couple options.
-                    # ++  We may have exhausted a particular parallel path
-                    #     from the current level, but there are more left. We
-                    #     need to refresh the list of instances and continue.
-                    # ++  We may have exhausted all subtrees of the current
-                    #     level. In that case, we need to back out a level and
-                    #     continue looking for parallels, one level up.
-                    if not instance_backlog_stack[-1]:
-                        # Note that by checking for >1, we don't allocate a
-                        # bunch of instance backlog children for nothing.
-                        if len(parallel_slot_backlog_stack[-1]) > 1:
-                            parallel_slot_backlog_stack[-1].pop()
-                            instance_backlog_stack[-1].extend(
-                                instance_history_stack[-1])
-
-                        else:
-                            parallel_slot_backlog_stack.pop()
-                            instance_backlog_stack.pop()
-                            instance_history_stack.pop()
-
-                    # There are one or more remaining parallel paths from the
-                    # current instances that lead to the target template_class.
-                    # Choose the last one so we can pop it efficiently.
-                    else:
-                        (
-                            nested_slot_name,
-                            nested_slot_type,
-                            nested_slot_routes
-                        ) = parallel_slot_backlog_stack[-1][-1]
-
-                        current_instance = (
-                            instance_backlog_stack[-1][0].instance)
-                        nested_instances = getattr(
-                            current_instance, nested_slot_name)
-                        nested_index = itertools.count()
-
-                        # The parallel path we chose is a leaf node,
-                        # which means that each nested instance is a
-                        # provenance. Note that, because of recursion loops,
-                        # this can happen whether or not there are nested
-                        # slot routes, so we'll still check for those in just
-                        # a second.
-                        if nested_slot_routes.is_terminus:
-                            partial_provenance = tuple(
-                                instance_backlog_level[0]
-                                for instance_backlog_level
-                                in instance_backlog_stack)
-
-                            # Note that using extend here is basically just a
-                            # shorthand for repeatedly iterating on the
-                            # outermost while loop after appending the
-                            # children (like we did with nested_slot_routes)
-                            provenances.extend(TemplateProvenance(
-                                (
-                                    *partial_provenance,
-                                    TemplateProvenanceNode(
-                                        encloser_slot_key=nested_slot_name,
-                                        encloser_slot_index=next(nested_index),
-                                        instance_id=id(nested_instance),
-                                        instance=nested_instance)))
-                                for nested_instance
-                                in nested_instances)
-
-                        # The parallel path we chose has more steps on the way
-                        # to the leaf node, so we need to continue deeper into
-                        # the tree.
-                        if nested_slot_routes:
-                            nested_provenances = tuple(
-                                TemplateProvenanceNode(
-                                    encloser_slot_key=nested_slot_name,
-                                    encloser_slot_index=next(nested_index),
-                                    instance_id=id(nested_instance),
-                                    instance=nested_instance)
-                                for nested_instance in nested_instances
-                                if isinstance(
-                                    nested_instance, nested_slot_type))
-                            instance_history_stack.append(nested_provenances)
-                            instance_backlog_stack.append(
-                                deque(nested_provenances))
-                            parallel_slot_backlog_stack.append(
-                                list(nested_slot_routes))
-
-                        # If there aren't any more nested slot routes, then
-                        # we need to back out one level on the stack.
-                        # Presumably this is also a terminus, but that's
-                        # handled above.
-                        else:
-                            # Note that we already popped from the parallel
-                            instance_backlog_stack[-1].popleft()
-
-                # Oh the humanity, oh the combinatorics!
-                invocations.extend(itertools.product(
-                    provenances,
-                    itertools.chain.from_iterable(
-                        abstract_calls.values()
-                    )))
-
-        root_provenance = TemplateProvenance((root_provenance_node,))
-        root_template_class = type(root_template_instance)
-        invocations.extend(
-            (root_provenance, abstract_call)
-            for abstract_call
-            in itertools.chain.from_iterable(
-                template_preload[root_template_class].function_calls.values()
-            ))
-        return invocations
-
-    def resolve_forward_ref(
-            self,
-            lookup_key: ForwardRefLookupKey,
-            resolved_template_cls: TemplateClass
-            ) -> None:
-        """Notifies a dependent class (one that declared a slot as a
-        forward reference) that the reference is now available, thereby
-        causing it to resolve the forward ref and remove it from its
-        pending trees.
-        """
-        enclosing_template_cls = self.template_cls_ref()
-        if enclosing_template_cls is None:
-            raise RuntimeError(
-                'Template class was garbage collected before template '
-                + 'signature, and then signature asked to resolve forward '
-                + 'ref?!')
-
-        _update_encloser_with_trees_from_slot(
-            enclosing_template_cls,
-            self._slot_tree_lookup,
-            self._pending_ref_lookup,
-            self._forward_ref_lookup_key,
-            resolved_template_cls,
-            self._pending_ref_lookup.pop(lookup_key).pending_root_node,)
-
-        self.refresh_included_template_classes_snapshot()
-        self.refresh_pending_forward_ref_registration()
-
-    def stringify_all(self) -> str:
-        """This is a debug method that creates a prettified string
-        version of the entire slot tree lookup (pending and concrete).
-        """
-        to_join = []
-        to_join.append('Resolved (concrete) slots:')
-        for template_class, root_node in self._slot_tree_lookup.items():
-            to_join.append(f'  {template_class}')
-            to_join.append(root_node.stringify(depth=1))
-
-        to_join.append('Pending (forward reference) slots:')
-        for ref_lookup_key, container in self._pending_ref_lookup.items():
-            to_join.append(f'  {ref_lookup_key}')
-            to_join.append(container.pending_root_node.stringify(depth=1))
-
-        return '\n'.join(to_join)
-
-
 def _extract_template_class_locals() -> dict[str, Any] | None:
     """When templates are created from inside a closure (ex, during
     testing, where this is extremely common), we need access to the
@@ -1442,21 +329,6 @@ def _extract_template_class_locals() -> dict[str, Any] | None:
     upmodule_frame = _get_first_frame_from_other_module()
     if upmodule_frame is not None:
         return upmodule_frame.f_locals
-
-
-def _extract_frame_scope_id() -> int | None:
-    """When templates are created from inside a closure (ex, during
-    testing, where this is extremely common), and forward references
-    are used, we need a way to differentiate between identically-named
-    templates within different functions of the same module (or the
-    toplevel of the module).
-
-    We do this via a dedicated decorator/context manager,
-    ``anchor_closure_scope``, which creates a random value and assigns
-    it to the corresponding context var, and then is retrieved by this
-    function for use.
-    """
-    return _CURRENT_SCOPE_ID.get()
 
 
 def _get_first_frame_from_other_module() -> FrameType | None:
@@ -1514,1058 +386,6 @@ def _classify_interface_field_flavor(
         return None
 
 
-@dataclass(frozen=True, slots=True)
-class _PendingSlotTreeContainer:
-    """Remember that the point here is to eventually build a lookup from
-    a ``{type[template]: _SlotTreeNode}``. And we're dealing with
-    forward references to the ``type[template]``, meaning we don't have
-    a key to use for the slot tree lookup. End of story.
-
-    So what we're doing here instead, is constructing the slot tree as
-    best as we can, and keeping track of what nodes need to be populated
-    by the forward reference, once it is resolved.
-
-    When the forward ref is resolved, we can simple copy the tree into
-    all of the insertion nodes, and then store the pending slot tree
-    in the slot tree lookup using the resolved template class.
-    """
-    pending_slot_type: ForwardRefLookupKey
-    pending_root_node: _PendingSlotTreeNode
-
-
-class ForwardReferenceProxyClass(Protocol):
-    REFERENCE_TARGET: ClassVar[ForwardRefLookupKey]
-
-
-def is_forward_reference_proxy(
-        obj: object
-        ) -> TypeIs[type[ForwardReferenceProxyClass]]:
-    return isinstance(obj, type) and hasattr(obj, 'REFERENCE_TARGET')
-
-
-@overload
-def _copy_slot_tree[T: _SlotTreeNode](
-        src_tree: T,
-        ) -> T: ...
-@overload
-def _copy_slot_tree[T: _SlotTreeNode](
-        src_tree: _SlotTreeNode,
-        *,
-        with_node_type: type[T],
-        ) -> T: ...
-def _copy_slot_tree[ST: _SlotTreeNode, NT: _SlotTreeNode](
-        src_tree: ST,
-        *,
-        with_node_type: type[NT] | None = None
-        ) -> ST | NT:
-    """This creates a copy of an existing slot tree. We use it when
-    merging nested slot trees into enclosers; otherwise, we end up with
-    a huge mess of "it's not clear what object holds which slot tree"
-    that is very difficult to reason about. This is slightly more memory
-    intensive, but... again, this is much, much easier to reason about.
-
-    Take special note that this preserves reference cycles, which is a
-    bit of a tricky thing.
-
-    Note: if ``into_tree`` is provided, this copies inplace and returns
-    the ``into_tree``. Otherwise, a new tree is created and returned.
-    In both cases, we also return a lookup from
-    ``{old_node.id_: copied_node}``.
-    """
-    copied_tree: ST | NT
-    if with_node_type is None:
-        node_type = type(src_tree)
-    else:
-        node_type = with_node_type
-
-    copied_tree = node_type()
-    copied_tree.merge_fields_only(src_tree)
-
-    # This converts ``old_node.id_`` to the new node instance; it's how we
-    # implement copying reference cycles
-    transmogrified_nodes: dict[int, ST | NT] = {src_tree.id_: copied_tree}
-    copy_stack: \
-        list[_SlotTreeTraversalFrame[ST | NT, ST]] = [
-            _SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=copied_tree,
-                insertion_subtree=src_tree,
-                first_encounters={})]
-
-    while copy_stack:
-        current_stack_frame = copy_stack[-1]
-        if current_stack_frame.exhausted:
-            copy_stack.pop()
-            continue
-
-        next_slot_route = current_stack_frame.insertion_subtree[
-            current_stack_frame.next_subtree_index]
-        next_slot_name, next_slot_type, next_subtree = next_slot_route
-        # Do this ASAP so that we don't accidentally forget it somehow
-        current_stack_frame.next_subtree_index += 1
-
-        next_subtree_id = next_subtree.id_
-        already_copied_node = transmogrified_nodes.get(next_subtree_id)
-        # This could be either the first time we hit a recursive subtree,
-        # or a non-recursive subtree.
-        if already_copied_node is None:
-            new_subtree = node_type()
-            new_subtree.merge_fields_only(next_subtree)
-
-            if next_subtree.requires_transmogrification:
-                transmogrified_nodes[next_subtree_id] = new_subtree
-
-            current_stack_frame.existing_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    new_subtree,))
-            copy_stack.append(_SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=new_subtree,
-                insertion_subtree=next_subtree,
-                # Note that, though this isn't correct, it also
-                # doesn't matter for purely COPYING a slot tree.
-                first_encounters={}))
-
-        # We've hit a recursive subtree -- one that we've already copied --
-        # which means we don't need to copy it again; instead, we just need to
-        # transmogrify the reference so that the nested route refers back to
-        # the original copied node.
-        else:
-            current_stack_frame.existing_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    already_copied_node,))
-
-    return copied_tree
-
-
-def _merge_into_slot_tree[T: _SlotTreeNode](
-        existing_tree_template_cls: TemplateClass,
-        # Note: this is only used for determining ``is_terminus``.
-        # Also note: None for pending slots. These can never be a terminus
-        # anyways, so it doesn't matter.
-        existing_tree_slot_type: TemplateClass | None,
-        existing_tree: T,
-        to_merge: T
-        ) -> None:
-    """This traverses the existing tree, merging in the slot_name and
-    its subtrees into the correct locations in the existing slot tree,
-    recursively.
-
-    This will cull any recursion loops to their minimum possible size.
-    Note that this can result in different slot tree recursion loop
-    "phases", based on which of the two slot types is encountered first.
-
-    Also note that this merges all public slot tree node attributes,
-    with three exceptions:
-    ++  the ``id_`` of the existing tree is always preserved
-    ++  ``is_recursive`` will always be calculated based on the
-        recursion loop culling
-    ++  ``is_terminus`` will always be calculated based on node types
-        matching the passed ``existing_tree_slot_type``.
-    """
-    node_type = type(existing_tree)
-    transmogrified_nodes: dict[int, T] = {to_merge.id_: existing_tree}
-    # Note: we want our internal culling logic to handle ``is_recursive``.
-    # Also note: the existing tree usually knows whether or not it's already
-    # recursive, but there's an edge case where it doesn't:
-    # 1.. have a recursive loop between two templates
-    # 2.. when resolving the back-ref of the second template, it finds a
-    #     forward ref to itself
-    # 3.. it now immediately resolves the forward ref to itself
-    # In this situation, the short-circuited self-referential forward ref
-    # results in is_terminus being False, which we correct here.
-    existing_tree.merge_fields_only(
-        to_merge, fields_to_skip={'is_recursive', 'is_terminus'})
-    existing_tree.is_terminus = (
-        existing_tree_template_cls is existing_tree_slot_type)
-
-    # Counterintuitive: since we're MERGING trees, the existing_subtree is
-    # actually the DESTINATION, and the insertion_subtree the source!
-    merge_stack: list[_SlotTreeTraversalFrame[T, T]] = [
-        _SlotTreeTraversalFrame(
-            next_subtree_index=0,
-            existing_subtree=existing_tree,
-            insertion_subtree=to_merge,
-            first_encounters={existing_tree_template_cls: existing_tree})]
-    # Yes, in theory, this one specific operation of merging trees would be
-    # faster if the trees were dicts instead of iterative structures. But
-    # we're not optimizing for tree merging; we're optimizing for rendering!
-    # And in that case, we're better off with a simple iterative structure.
-    while merge_stack:
-        current_stack_frame = merge_stack[-1]
-        if current_stack_frame.exhausted:
-            merge_stack.pop()
-            continue
-
-        existing_subtree = current_stack_frame.existing_subtree
-        next_slot_route = current_stack_frame.insertion_subtree[
-            current_stack_frame.next_subtree_index]
-        next_slot_name, next_slot_type, next_subtree = next_slot_route
-        # Do this ASAP so that we don't accidentally forget it somehow (also
-        # because we want to use a continue statement in a second)
-        current_stack_frame.next_subtree_index += 1
-
-        next_subtree_id = next_subtree.id_
-        already_merged_node = transmogrified_nodes.get(next_subtree_id)
-
-        # For merging, we're going to handle the recursive subtree case first,
-        # because it makes the rest of the logic cleaner
-        if already_merged_node is not None:
-            # Note that appending is idempotent as long as the
-            # already_merged_node is the same object as any hypothetical
-            # already_existing_already_merged_node that could be found there.
-            # So this will only error if the slot path already exists, AND
-            # it points to a different node.
-            current_stack_frame.existing_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name, next_slot_type, already_merged_node))
-
-            already_merged_node.merge_fields_only(
-                next_subtree,
-                # Note: we want our internal culling logic to handle
-                # ``is_recursive``. Otherwise, we might result in something
-                # being reported as recursive, which actually isn't -- because
-                # the ... phase, I guess you could say ... of the recursion
-                # loop is different with a different root node.
-                # Also note: we don't want to preserve the terminus value of
-                # a tree we're merging in; we want the terminus to be purely
-                # calculated within the merge function. This prevents bugs
-                # when merging trees across slot types, which happens a lot
-                # with pending slot trees.
-                fields_to_skip={'is_recursive', 'is_terminus'})
-            # This might seem redundant, but there are some weird edge cases
-            # when merging a fully-defined tree (for example, adding in
-            # self-referential recursion loops to a pending tree) that require
-            # it to be set.
-            # (This is because we skip it when merging the fields on the
-            # existing tree; this recovers the correct state).
-            already_merged_node.is_recursive = True
-            continue
-
-        # This accomplishes two things: first, it culls an extra cycle from the
-        # to_merge tree that would ultimately have the same effect. Secondly,
-        # it ensures correct recursion when we're merging in the pending tree
-        # from a class that used us as a forward reference, since the other
-        # class won't yet be resolved.
-        # Note that, by definition, this won't happen the first time we
-        # encounter a node of any particular type, including the type of the
-        # rootwards-most node for the whole tree. This is expected! We first
-        # need to encounter the slot type before we can recurse to it. **We
-        # will never recurse back to the root of the slot tree!**
-        if next_slot_type in current_stack_frame.first_encounters:
-            # Note that we want to avoid modifying the already-encountered
-            # tree. This is mostly because we copy trees willy nilly and want
-            # to avoid them accidentally getting out of sync. However, it does
-            # require us to be very careful to ensure that we've corrently
-            # copied all needed values over prior to merging trees, if we've
-            # just created the existing tree from scratch.
-            next_existing_subtree = current_stack_frame.first_encounters[
-                next_slot_type]
-            next_existing_subtree.is_recursive = True
-
-            next_existing_route = _SlotTreeRoute.new(
-                next_slot_name,
-                next_slot_type,
-                next_existing_subtree)
-            if existing_subtree.has_route_for(next_slot_name, next_slot_type):
-                existing_subtree.rewrite_route_for(
-                    next_slot_name, next_slot_type, next_existing_route)
-                # raise RecursionError(
-                #     'Non-culled infinite recursion while merging templatey '
-                #     + 'slot trees!', next_slot_name, next_slot_type)
-
-            else:
-                existing_subtree.append(next_existing_route)
-
-            # Note that we don't need to update transmogrification for two
-            # reasons: first, because we added the root node at the very
-            # beginning, and second, because we -- by definition -- cannot
-            # have any deeper references to this part of the destination tree,
-            # because  we're culling the depthwise-rest of the source tree
-            continue
-
-        # The existing subtree -- the one we're merging INTO -- has a route for
-        # this slot name and type already, so we need to merge them together
-        # instead of simply copy/transmogrify/cull
-        elif existing_subtree.has_route_for(next_slot_name, next_slot_type):
-            next_existing_route = existing_subtree.get_route_for(
-                    next_slot_name, next_slot_type)
-            __, __, next_existing_subtree = next_existing_route
-            next_existing_subtree.merge_fields_only(
-                next_subtree,
-                # Note: we want our internal culling logic to handle
-                # ``is_recursive``. Otherwise, we might result in something
-                # being reported as recursive, which actually isn't -- because
-                # the ... phase, I guess you could say ... of the recursion
-                # loop is different with a different root node.
-                # Also note: we don't want to preserve the terminus value of
-                # a tree we're merging in; we want the terminus to be purely
-                # calculated within the merge function. This prevents bugs
-                # when merging trees across slot types, which happens a lot
-                # with pending slot trees.
-                fields_to_skip={'is_recursive', 'is_terminus'})
-
-        # The existing subtree doesn't have any existing routes for this, so
-        # we don't need to worry about merging things together -- but we still
-        # need to worry about transmogrification and culling.
-        # Also note that there might be identically-named slots for different
-        # slot types in the case of a union, but that will be handled on a
-        # different iteration of the merge stack while loop.
-        else:
-            # Note that the next_subtree might have a different node type
-            # than the existing tree!
-            next_existing_subtree = node_type()
-            next_existing_subtree.merge_fields_only(
-                next_subtree,
-                fields_to_skip={'is_recursive', 'is_terminus'})
-            # Because we do lots and lots of merging, with offsets and
-            # transforms etc, it's easiest to just reset this every time.
-            # This helps prevent weird bugs with copying.
-            next_existing_subtree.is_terminus = (
-                # Note: what matters here is the slot type, not the template
-                # class!
-                next_slot_type is existing_tree_slot_type)
-            next_existing_route = _SlotTreeRoute.new(
-                next_slot_name,
-                next_slot_type,
-                next_existing_subtree)
-            existing_subtree.append(next_existing_route)
-
-        if next_existing_subtree.requires_transmogrification:
-            transmogrified_nodes[next_subtree_id] = next_existing_subtree
-
-        if next_subtree:
-            next_first_encounters = {**current_stack_frame.first_encounters}
-            next_first_encounters.setdefault(
-                next_slot_type, next_existing_subtree)
-            merge_stack.append(_SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=next_existing_subtree,
-                insertion_subtree=next_subtree,
-                first_encounters=next_first_encounters))
-
-
-def _update_encloser_with_trees_from_slot(
-        enclosing_cls: TemplateClass,
-        enclosing_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        enclosing_pending_ref_lookup:
-            dict[ForwardRefLookupKey, _PendingSlotTreeContainer],
-        enclosing_cls_ref_lookup_key: ForwardRefLookupKey,
-        nested_slot_type: TemplateClass,
-        nested_slot_offset: _PendingSlotTreeNode,
-        ) -> None:
-    """This function updates the enclosing class' concrete and pending
-    slot trees with the concrete and pending slot trees from the
-    nested slot.
-
-    Note that it can be called during initial signature assembly, hence
-    needing explicit arguments instead of relying upon the lookups
-    already being defined on the class.
-    """
-    nested_slot_xable = cast(type[TemplateIntersectable], nested_slot_type)
-    nested_lookup = nested_slot_xable._templatey_signature._slot_tree_lookup
-    nested_pending_refs = (
-        nested_slot_xable._templatey_signature._pending_ref_lookup)
-    # First and foremost, we can't forget to add a route for the
-    # actual slot itself!
-    _apply_concrete_insertions(
-        enclosing_cls,
-        enclosing_slot_tree_lookup,
-        nested_slot_type,
-        nested_lookup,
-        nested_slot_type,
-        # Note that this is actually supposed to be the nested slot offset
-        # in both cases; the function itself gets the appropriate tree from
-        # the nested_lookup
-        nested_slot_offset)
-
-    # Now the CONCRETE nested slots.
-    # IMPORTANT: all of the nested concrete slots need to already
-    # be merged in before starting the nested pending refs! (More
-    # info below).
-    for doubly_nested_slot_type in nested_lookup:
-        _apply_concrete_insertions(
-            enclosing_cls,
-            enclosing_slot_tree_lookup,
-            nested_slot_type,
-            nested_lookup,
-            doubly_nested_slot_type,
-            # Note that this is actually supposed to be the nested slot offset
-            # in both cases; the function itself gets the appropriate tree from
-            # the nested_lookup
-            nested_slot_offset)
-
-    # And now, finally, the nested PENDING slots.
-    # IMPORTANT: all of the nested concrete slots need to already
-    # be merged in before starting the nested pending refs!
-    # If we reversed the order, then we wouldn't be able to
-    # guarantee that we'd already discovered every possible
-    # terminus node for the nested slot type (due to potential
-    # recursion loops). This could then cause us to miss out on
-    # a lot of the combinatorics of the offset pending trees.
-    # That being said, note that this is just a transform-and-merge
-    # operation; we're not adding any additional insertion points
-    # on top of what the nested slot defines. That happens later,
-    # when we deal with the pending ref defs for the enclosing
-    # template class (the one we're currently constructing a
-    # signature for).
-
-    # Note: we want to special-case a nested forward ref to the
-    # enclosing class (ie, a recursion loop) for exactly the
-    # same reason: to guarantee it's resolved before we start on
-    # the other nested ones. Hence copy-and-mutate.
-    unresolved_nested_forward_refs = {**nested_pending_refs}
-
-    # Remember that we're potentially in the middle of constructing the
-    # signature for a new template class. If the nested class
-    # (from the slot) was depending on the class we're still
-    # constructing, it hasn't yet been updated with the
-    # resolved class. Therefore, instead of needing to come
-    # back and fix up any recursive forward refs later, we can
-    # simply do them right here, right now.
-    # Also note that we'll NEVER have an existing pending tree
-    # for this, because we're adding it directly and
-    # immediately into the actual slot tree.
-    encloser_referencing_pending_container = (
-        unresolved_nested_forward_refs.pop(
-            enclosing_cls_ref_lookup_key, None))
-    if encloser_referencing_pending_container is not None:
-        # This is effectively a special-case of pending slot tree extension
-        # where we immediately apply the concrete insertions. But we still need
-        # to merge the two pending trees (which handles the combinatorics --
-        # remember that each insertion point for the nested class will also
-        # get N insertion points from its references to the enclosing cls).
-        _extend_pending_slot_tree(
-            enclosing_cls=enclosing_cls,
-            enclosing_slot_tree_lookup=enclosing_slot_tree_lookup,
-            enclosing_pending_ref_lookup=enclosing_pending_ref_lookup,
-            nested_cls=nested_slot_type,
-            nested_pending_container=encloser_referencing_pending_container,
-            nested_pending_ref=enclosing_cls_ref_lookup_key,)
-        offset_pending_container = enclosing_pending_ref_lookup.pop(
-            enclosing_cls_ref_lookup_key)
-        _apply_concrete_insertions(
-            enclosing_cls,
-            enclosing_slot_tree_lookup,
-            # These are counter-intuitive, but correct. We've divorced this
-            # from the nested slot entirely; all the insertion points for it
-            # now are pointed back at ourselves.
-            enclosing_cls,
-            enclosing_slot_tree_lookup,
-            enclosing_cls,
-            offset_pending_container.pending_root_node)
-
-    # Okay, after all of that meticulous work, we can now guarantee
-    # that all of the concrete classes we know about up until this
-    # point have been fully resolved and incorporated into the
-    # slot tree. That means we can FINALLY start merging the
-    # pending items from the nested slot.
-    for (
-        nested_forward_ref_key, nested_pending_container
-    ) in unresolved_nested_forward_refs.items():
-        _extend_pending_slot_tree(
-            enclosing_cls=enclosing_cls,
-            enclosing_slot_tree_lookup=enclosing_slot_tree_lookup,
-            enclosing_pending_ref_lookup=enclosing_pending_ref_lookup,
-            nested_cls=nested_slot_type,
-            nested_pending_container=nested_pending_container,
-            nested_pending_ref=nested_forward_ref_key,)
-
-
-def _resolve_pending_slot_tree(  # noqa: PLR0913
-        pending_tree: _PendingSlotTreeContainer,
-        slot_info: tuple[str, TemplateClass] | None,
-        enclosing_cls: TemplateClass,
-        enclosing_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        enclosing_pending_ref_lookup:
-            dict[ForwardRefLookupKey, _PendingSlotTreeContainer],
-        resolved_cls: TemplateClass,
-        resolved_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        resolved_pending_ref_lookup:
-            dict[ForwardRefLookupKey, _PendingSlotTreeContainer],
-        ) -> None:
-    """Once a forward ref has been resolved, it needs to have both the
-    concrete and pending trees from the now-resolved class merged into
-    the respective trees of the enclosing class.
-
-    There are several things to be mindful of here:
-    ++  Pending trees are really factories for concrete slot trees. Each
-        nested template class in the resolved class will be inserted
-        at the pending tree insertion point. In other words, if there
-        are N insertion points, and the resolved class has M slots, then
-        N * M concretified slot trees will need to be merged into the
-        enclosing class.
-    ++  The slot tree of the resolved class (or its pending tree) may
-        include references to the enclosing class. These subtrees need
-        to then be short-circuited, and replaced by references to the
-        slot tree root for the enclosing class.
-    ++  The resolved class may not be completely finalized yet, since
-        we are also called during initial tree creation for the resolved
-        class (though only when recursive loops are encountered).
-        Therefore, we need to explicitly pass the concrete and pending
-        slot tree lookups for each of the enclosing and resolved
-        classes, in addition to the class itself
-    ++  The resolved and enclosing classes might be the same class (as
-        in the above case, where this is called during initial tree
-        creation for a resolved class in a recursive loop).
-    """
-    pending_tree_root_node = pending_tree.pending_root_node
-    if slot_info is None:
-        # Note: the root of the tree is only a terminus if the slot type
-        # matches the enclosing class, and slot_info=None
-        # implies that we're working from the root of the tree.
-        resolved_slot_item = (resolved_cls, _SlotTreeNode())
-
-        # Note: again, we're resolving against the root tree. That
-        # means that the pending tree is relative to it, and not to a nested
-        # slot. This also implies that we're resolving a forward ref on an
-        # already-existing template class.
-        # It also means we don't need to offset the given tree.
-        offset_pending_tree_root_node = pending_tree_root_node
-
-    else:
-        slot_name, slot_cls = slot_info
-        resolved_slot_item = (
-            resolved_cls,
-            _SlotTreeNode([_SlotTreeRoute.new(
-                slot_name, slot_cls, _SlotTreeNode())]))
-
-        # However, in this case, we're resolving against a specific slot on
-        # a root key. This implies that we're resolving a forward ref on a
-        # not-yet-exited-for-the-decorator template class.
-        # This means the naked tree isn't yet nested under the correct slot.
-        # Note that the nesting is crucial; that's what puts the tree under
-        # the correct slot for the enclosing class, instead of retaining
-        # the tree depth and root slot from the resolved class.
-        offset_pending_tree_root_node = _PendingSlotTreeNode([
-            _SlotTreeRoute.new(
-                slot_name,
-                slot_cls,
-                pending_tree_root_node)])
-
-    # We need to make sure to include the actual resolved class itself in
-    # the slot tree! But instead of special-casing it, we can just let it be
-    # handled by the same logic for the nested ones.
-    all_nested_slot_items = itertools.chain(
-        resolved_slot_tree_lookup.items(),
-        [resolved_slot_item])
-    for nested_template_cls, nested_root_node in all_nested_slot_items:
-        # We're about to change some of the fields here based on the nested
-        # template class, so we want to make sure we're only modifying a copy.
-        copied_offset_pending_tree_root_node = _copy_slot_tree(
-            offset_pending_tree_root_node)
-
-        # Note that, even though we're iterating over mostly just the
-        # resolved_slot_tree_lookup.items(), we've added the resolved_cls to
-        # the mix, which may or may not be in the lookup; therefore, this check
-        # is necessary.
-        if nested_template_cls in resolved_slot_tree_lookup:
-            root_node = resolved_slot_tree_lookup[nested_template_cls]
-            copied_offset_pending_tree_root_node.merge_fields_only(root_node)
-
-        # Note: it's important to do the offsets BEFORE applying insertions,
-        # because it allows recursion detection to function correctly against
-        # the root node.
-        tree_after_insertions = _apply_insertions(
-            resolved_cls,
-            copied_offset_pending_tree_root_node,
-            nested_template_cls,
-            nested_root_node)
-        # The root node of a slot tree pointing back at the enclosing class --
-        # ie, a recursive slot -- must have is_terminus set to True, otherwise
-        # recursion won't work correctly, since the recursive nodes will be
-        # pointing at the root object.
-        tree_after_insertions.is_terminus |= (
-            nested_template_cls is enclosing_cls)
-
-        dest_tree = enclosing_slot_tree_lookup.get(nested_template_cls)
-        # Note that we still need to call merge_into to perform any needed
-        # culling, so we can't simply assign the dest_tree as the value in the
-        # enclosing lookup
-        if dest_tree is None:
-            dest_tree = enclosing_slot_tree_lookup[nested_template_cls] = (
-                # Copying over the ID isn't necessary, but it makes the
-                # lineage a bit clearer if you need to debug things
-                _SlotTreeNode(id_=tree_after_insertions.id_))
-
-        # This, however, IS necessary. The destination tree is used as the
-        # first encounter of the template class, and merging avoids
-        # modifying already-encountered nodes. Therefore, we have to
-        # explicitly merge the values of tree_after_insertions to avoid
-        # discarding them!
-        dest_tree.merge_fields_only(tree_after_insertions)
-
-        _merge_into_slot_tree(
-            enclosing_cls,
-            nested_template_cls,
-            dest_tree,
-            tree_after_insertions)
-
-    for (
-        forward_ref_key,
-        pending_tree_container
-    ) in resolved_pending_ref_lookup.items():
-        # We're about to change some of the fields here based on the nested
-        # template class, so we want to make sure we're only modifying a copy.
-        copied_offset_pending_tree_root_node = _copy_slot_tree(
-            offset_pending_tree_root_node)
-        # Note that -- unlike with the concrete tree -- because we're not
-        # injecting anything here, this is always correct.
-        copied_offset_pending_tree_root_node.merge_fields_only(
-            pending_tree_container.pending_root_node)
-
-        # Note: it's important to do the offsets BEFORE applying insertions,
-        # because it allows recursion detection to function correctly against
-        # the root node.
-        tree_after_insertions = _apply_insertions(
-            resolved_cls,
-            copied_offset_pending_tree_root_node,
-            forward_ref_key,
-            pending_tree_container.pending_root_node)
-
-        dest_pending = enclosing_pending_ref_lookup.get(forward_ref_key)
-        # Note that we still need to call merge_into to perform any needed
-        # culling, so we can't simply assign the dest_tree as the value in the
-        # enclosing lookup
-        if dest_pending is None:
-            # Copying over the ID isn't necessary, but it makes the
-            # lineage a bit clearer if you need to debug things
-            dest_tree =  _PendingSlotTreeNode(id_=tree_after_insertions.id_)
-            enclosing_pending_ref_lookup[forward_ref_key] = (
-                _PendingSlotTreeContainer(forward_ref_key, dest_tree))
-        else:
-            dest_tree = dest_pending.pending_root_node
-
-        # This, however, IS necessary, and especially important for the
-        # insertion slot names!
-        dest_tree.merge_fields_only(tree_after_insertions)
-
-        _merge_into_slot_tree(
-            enclosing_cls,
-            None,
-            dest_tree,
-            tree_after_insertions)
-
-
-def _apply_concrete_insertions(
-        enclosing_cls: TemplateClass,
-        enclosing_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        nested_slot_type: TemplateClass,
-        # Note: may or may not be different from insertion_cls; see note in
-        # docstring
-        nested_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        insertion_slot_type: TemplateClass,
-        insertion_tree: _PendingSlotTreeNode
-        ) -> None:
-    """This takes an insertion tree (a pending slot tree node), applies
-    a resolved, concrete insertion class at each of the insertion
-    points on the insertion tree, and then merges the result into the
-    slot tree lookup for the enclosing class.
-
-    Note that this is meant to be used both with the nested class
-    itself, as well as its (doubly-)nested concrete slots. In the former
-    case, the slot tree lookup will match the insertion class. In the
-    latter case, the insertion class will instead be the (doubly-)nested
-    concrete slot.
-    """
-    root_after_insertion = _SlotTreeNode()
-    tree_to_insert = nested_slot_tree_lookup.get(
-        insertion_slot_type,
-        _SlotTreeNode(is_terminus=True))
-
-    # This converts ``old_node.id_`` to the new node instance; it's how we
-    # implement copying reference cycles
-    transmogrified_nodes: dict[int, _SlotTreeNode] = {
-        insertion_tree.id_: root_after_insertion}
-    stack: \
-        list[_SlotTreeTraversalFrame[_SlotTreeNode, _PendingSlotTreeNode]] = [
-        _SlotTreeTraversalFrame(
-            next_subtree_index=0,
-            existing_subtree=root_after_insertion,
-            insertion_subtree=insertion_tree,
-            first_encounters={enclosing_cls: root_after_insertion})]
-
-    while stack:
-        current_stack_frame = stack[-1]
-        src_subtree = current_stack_frame.insertion_subtree
-        target_subtree = current_stack_frame.existing_subtree
-
-        if current_stack_frame.exhausted:
-            # This is counter-intuitive. Note that:
-            # ++  the stack is for checking nested (deeper) routes only. it
-            #     does not check for insertions on the current stack frame.
-            # ++  the current stack frame might, though, actually have some
-            #     insertions!
-            # ++  we only want those insertions to be applied ONCE, and not
-            #     once per subtree
-            # ++  ordering of the insertion application doesn't matter
-            # Therefore, we apply insertions to the current frame immediately
-            # before discarding the frame for being exhausted.
-            for nested_slot_type_slot_name in src_subtree.insertion_slot_names:
-                target_subtree.append(
-                    _SlotTreeRoute.new(
-                        nested_slot_type_slot_name,
-                        nested_slot_type,
-                        _copy_slot_tree(tree_to_insert)))
-
-            stack.pop()
-            continue
-
-        next_slot_route = src_subtree[current_stack_frame.next_subtree_index]
-        next_slot_name, next_slot_type, next_subtree = next_slot_route
-        # Do this ASAP so that we don't accidentally forget it somehow
-        current_stack_frame.next_subtree_index += 1
-
-        # Note that this will still get merged into the actual full slot tree
-        # for the enclosing template, which will cull any extra links in
-        # recursive reference cycles, so we don't need to worry about that
-        # here.
-        next_subtree_id = next_subtree.id_
-        transmogrified_dest_subtree = transmogrified_nodes.get(next_subtree_id)
-
-        # This could be either the first time we hit a recursive subtree,
-        # or a non-recursive subtree.
-        if transmogrified_dest_subtree is None:
-            dest_subtree = _SlotTreeNode()
-            dest_subtree.merge_fields_only(next_subtree)
-
-            if next_subtree.requires_transmogrification:
-                transmogrified_nodes[next_subtree_id] = dest_subtree
-
-            # Note that, since we're building a new tree from scratch, we don't
-            # need to worry about this already existing.
-            target_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    dest_subtree))
-            next_first_encounters = {**current_stack_frame.first_encounters}
-            next_first_encounters.setdefault(next_slot_type, dest_subtree)
-            stack.append(_SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=dest_subtree,
-                insertion_subtree=next_subtree,
-                first_encounters=next_first_encounters))
-
-        # We've hit a recursive subtree -- one that we've already copied --
-        # which means we don't need to copy it again; instead, we just need to
-        # transmogrify the reference so that the nested route refers back to
-        # the original copied node.
-        else:
-            dest_subtree = transmogrified_dest_subtree
-            # Note: is_recursive was already set!
-            target_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    transmogrified_dest_subtree))
-
-    enclosing_root = enclosing_slot_tree_lookup.get(insertion_slot_type)
-    if enclosing_root is None:
-        enclosing_root = enclosing_slot_tree_lookup[insertion_slot_type] = (
-            _SlotTreeNode())
-
-    _merge_into_slot_tree(
-        enclosing_cls,
-        insertion_slot_type,
-        enclosing_root,
-        root_after_insertion)
-
-
-@overload
-def _apply_insertions(
-        resolved_cls: TemplateClass,
-        pending_tree: _PendingSlotTreeNode,
-        nested_template_type: TemplateClass,
-        nested_root_node: _ConcreteSlotTreeNode
-        ) -> _ConcreteSlotTreeNode: ...
-@overload
-def _apply_insertions(
-        resolved_cls: TemplateClass,
-        pending_tree: _PendingSlotTreeNode,
-        nested_template_type: ForwardRefLookupKey,
-        nested_root_node: _PendingSlotTreeNode
-        ) -> _PendingSlotTreeNode: ...
-def _apply_insertions[T: _ConcreteSlotTreeNode | _PendingSlotTreeNode](
-        resolved_cls: TemplateClass,
-        pending_tree: _PendingSlotTreeNode,
-        nested_template_type: TemplateClass | ForwardRefLookupKey,
-        nested_root_node: T,
-        ) -> T:
-    """For a particular pending tree and a single nested CONCRETE root
-    node, this searches the pending tree for all insertion points and
-    then inserts the concrete root node at that point.
-
-    **Note that this does not cull any superfluous recursive
-    refererence cycles.** You still need to merge the resulting tree
-    into the actual slot tree lookup, and then that does the culling.
-    """
-    if isinstance(nested_template_type, ForwardRefLookupKey):
-        first_encounter_cls = None
-    else:
-        first_encounter_cls = nested_template_type
-
-    resulting_tree: T = type(nested_root_node)()
-    resulting_tree.merge_fields_only(pending_tree)
-    # This converts ``old_node.id_`` to the new node instance; it's how we
-    # implement copying reference cycles
-    transmogrified_nodes: dict[int, T] = {
-        pending_tree.id_: resulting_tree}
-
-    stack: \
-        list[_SlotTreeTraversalFrame[_SlotTreeNode, _PendingSlotTreeNode]] = [
-        _SlotTreeTraversalFrame(
-            next_subtree_index=0,
-            existing_subtree=resulting_tree,
-            insertion_subtree=pending_tree,
-            first_encounters={first_encounter_cls: resulting_tree})]
-
-    while stack:
-        current_stack_frame = stack[-1]
-        src_subtree = current_stack_frame.insertion_subtree
-        target_subtree = current_stack_frame.existing_subtree
-
-        if current_stack_frame.exhausted:
-            # This is counter-intuitive. Note that:
-            # ++  the stack is for checking nested (deeper) routes only. it
-            #     does not check for insertions on the current stack frame.
-            # ++  the current stack frame might, though, actually have some
-            #     insertions!
-            # ++  we only want those insertions to be applied ONCE, and not
-            #     once per subtree
-            # ++  ordering of the insertion application doesn't matter
-            # Therefore, we apply insertions to the current frame immediately
-            # before discarding the frame for being exhausted.
-            for insertion_slot_name in src_subtree.insertion_slot_names:
-                target_subtree.append(
-                    _SlotTreeRoute.new(
-                        insertion_slot_name,
-                        resolved_cls,
-                        _copy_slot_tree(nested_root_node)))
-
-            stack.pop()
-            continue
-
-        next_slot_route = src_subtree[current_stack_frame.next_subtree_index]
-        next_slot_name, next_slot_type, next_subtree = next_slot_route
-        # Do this ASAP so that we don't accidentally forget it somehow
-        current_stack_frame.next_subtree_index += 1
-
-        # Note that this will still get merged into the actual full slot tree
-        # for the enclosing template, which will cull any extra links in
-        # recursive reference cycles, so we don't need to worry about that
-        # here.
-        next_subtree_id = next_subtree.id_
-        transmogrified_dest_subtree = transmogrified_nodes.get(next_subtree_id)
-
-        # This could be either the first time we hit a recursive subtree,
-        # or a non-recursive subtree.
-        if transmogrified_dest_subtree is None:
-            dest_subtree = type(nested_root_node)()
-            dest_subtree.merge_fields_only(next_subtree)
-
-            if next_subtree.requires_transmogrification:
-                transmogrified_nodes[next_subtree_id] = dest_subtree
-
-            target_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    dest_subtree))
-            next_first_encounters = {**current_stack_frame.first_encounters}
-            next_first_encounters.setdefault(next_slot_type, dest_subtree)
-            stack.append(_SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=dest_subtree,
-                insertion_subtree=next_subtree,
-                first_encounters=next_first_encounters))
-
-        # We've hit a recursive subtree -- one that we've already copied --
-        # which means we don't need to copy it again; instead, we just need to
-        # transmogrify the reference so that the nested route refers back to
-        # the original copied node.
-        else:
-            dest_subtree = transmogrified_dest_subtree
-            # Note: is_recursive was already set!
-            target_subtree.append(
-                _SlotTreeRoute.new(
-                    next_slot_name,
-                    next_slot_type,
-                    transmogrified_dest_subtree))
-
-    return resulting_tree
-
-
-def _extend_pending_slot_tree(
-        enclosing_cls: TemplateClass,
-        enclosing_slot_tree_lookup: dict[TemplateClass, _ConcreteSlotTreeNode],
-        enclosing_pending_ref_lookup:
-            dict[ForwardRefLookupKey, _PendingSlotTreeContainer],
-        nested_cls: TemplateClass,
-        nested_pending_container: _PendingSlotTreeContainer,
-        nested_pending_ref: ForwardRefLookupKey,
-        ) -> None:
-    """When we resolve a pending class, we do (of course) resolve the
-    pending class itself. But in the process, we may also suddenly have
-    a bunch of new pending refs, courtesy of the pending class.
-
-    At first glance, you might think that we could simply dump these
-    into the enclosing pending ref lookup, offset by the slot path for
-    the newly-resolved class. HOWEVER, this ignores recursion, and is
-    fragile in situations where classes have multiple slot names for
-    a particular slot type.
-
-    Instead, what we really need to do, is:
-    ++  create a pending tree copy of the concrete slot tree for the
-        newly-resolved class
-    ++  at every terminus:
-        ++  set ``is_terminus`` to False
-        ++  insert a copy of the nested root tree
-    """
-    copied_tree = _copy_slot_tree(
-        enclosing_slot_tree_lookup[nested_cls],
-        with_node_type=_PendingSlotTreeNode)
-    # Note that it's not an issue that this doesn't track the stack, because
-    # we aren't changing the structure of the existing tree at all.
-    # (We can't reuse first_encounters because it's meant for a TemplateClass)
-    encountered_node_ids: set[int] = set()
-
-    # It's a little weird to be repurposing the slot tree traversal frame
-    # when the existing and insertion subtrees are always the same, but it
-    # won't break anything, so we might as well.
-    stack: \
-        list[
-            _SlotTreeTraversalFrame[_PendingSlotTreeNode,
-            _PendingSlotTreeNode]] = [_SlotTreeTraversalFrame(
-                next_subtree_index=0,
-                existing_subtree=copied_tree,
-                insertion_subtree=copied_tree,
-                first_encounters={})]
-
-    while stack:
-        current_stack_frame = stack[-1]
-        if current_stack_frame.exhausted:
-            stack.pop()
-            continue
-
-        next_slot_route = current_stack_frame.insertion_subtree[
-            current_stack_frame.next_subtree_index]
-        __, next_slot_type, next_subtree = next_slot_route
-        # Do this ASAP so that we don't accidentally forget it somehow
-        current_stack_frame.next_subtree_index += 1
-
-        # There's never anything to do here; it means we've already done all
-        # of our transforms and insertions, because it's always recursive.
-        # Doesn't matter where on the tree we are. We've already fixed this
-        # node, period, end of story.
-        if next_subtree.id_ in encountered_node_ids:
-            continue
-        encountered_node_ids.add(next_subtree.id_)
-
-        if next_subtree.is_terminus:
-            # As per docstring, the idea here is that we're converting every
-            # terminus into an insertion point for a different (pending) slot.
-            # Therefore it's no longer a terminus, since the slot is different.
-            next_subtree.is_terminus = False
-            _merge_into_slot_tree(
-                # Note: this is intentionally NOT the enclosing_cls! We're
-                # working on an offset tree here, and we need to make sure that
-                # we correctly report the class of the OFFSET root!
-                next_slot_type,
-                None,
-                next_subtree,
-                nested_pending_container.pending_root_node)
-
-            # Also note that _merge_into_slot_tree assumes that the root node
-            # it's given is already correct. Therefore we need to manually
-            # update it here, in case there were attributes defined on it
-            # directly, and not as nested nodes.
-            next_subtree.merge_fields_only(
-                nested_pending_container.pending_root_node)
-
-    # Okay, the transform is complete; now we just need to insert it into
-    # the actual pending slot tree for the enclosing template class.
-    if nested_pending_ref in enclosing_pending_ref_lookup:
-        dest_pending_tree_container = enclosing_pending_ref_lookup[
-            nested_pending_ref]
-    else:
-        dest_pending_tree_container = enclosing_pending_ref_lookup[
-            nested_pending_ref] = _PendingSlotTreeContainer(
-                pending_slot_type=nested_pending_ref,
-                pending_root_node=_PendingSlotTreeNode())
-
-    _merge_into_slot_tree(
-        enclosing_cls,
-        None,
-        dest_pending_tree_container.pending_root_node,
-        copied_tree)
-
-
-@dataclass(slots=True)
-class _SlotTreeTraversalFrame[ET: _SlotTreeNode, IT: _SlotTreeNode]:
-    next_subtree_index: int
-    existing_subtree: ET
-    insertion_subtree: IT
-    first_encounters: dict[TemplateClass | None, ET]
-
-    @property
-    def exhausted(self) -> bool:
-        """Returns True if the to_merge_subtree has been exhausted, and
-        there are no more subtrees to merge.
-        """
-        return self.next_subtree_index >= len(self.insertion_subtree)
-
-
-# Note: mutablemapping because otherwise chainmap complains. Even though they
-# aren't actually implemented, this is a quick way of getting typing to work
-@dataclass(kw_only=True, slots=True)
-class _ForwardRefGeneratingNamespaceLookup(MutableMapping[str, type]):
-    template_module: str
-    template_scope_id: int | None
-    captured_refs: set[ForwardRefLookupKey] = field(default_factory=set)
-
-    def __getitem__(self, key: str) -> type:
-        forward_ref = ForwardRefLookupKey(
-            module=self.template_module,
-            name=key,
-            scope_id=self.template_scope_id)
-
-        class ForwardReferenceProxyClass:
-            """When we return a forward reference, we want to retain all
-            of the expected behavior with types -- unions via ``|``,
-            etc -- and therefore, we want to return a proxy class
-            instead of the forward reference itself.
-            """
-            REFERENCE_TARGET = forward_ref
-
-        self.captured_refs.add(forward_ref)
-        return ForwardReferenceProxyClass
-
-    # Required for mutable mapping protocol, but not for the namespace lookup.
-    def __iter__(self) -> Iterator[str]:
-        raise TypeError(
-            'Unsupported method call in templatey foward ref implementation.')
-
-    # Required for mutable mapping protocol, but not for the namespace lookup.
-    def __len__(self) -> int:
-        raise TypeError(
-            'Unsupported method call in templatey foward ref implementation.')
-
-    # Required for mutable mapping protocol, but not for the namespace lookup.
-    def __setitem__(self, key, value) -> None:
-        raise TypeError(
-            'Unsupported method call in templatey foward ref implementation.')
-
-    # Required for mutable mapping protocol, but not for the namespace lookup.
-    def __delitem__(self, key) -> None:
-        raise TypeError(
-            'Unsupported method call in templatey foward ref implementation.')
-
-
 @dataclass_transform(field_specifiers=(param, field, Field))
 def make_template_definition[T: type](
         cls: T,
@@ -2585,7 +405,7 @@ def make_template_definition[T: type](
     cls._templatey_resource_locator = template_resource_locator
 
     template_module = cls.__module__
-    template_scope_id = _extract_frame_scope_id()
+    template_scope_id = extract_frame_scope_id()
     template_forward_ref = ForwardRefLookupKey(
         module=template_module,
         name=cls.__name__,
@@ -2655,6 +475,7 @@ def make_template_definition[T: type](
     vars_ = {}
     content = {}
     prerenderers = {}
+    # Note that this ignores initvars, which is what we want
     for template_field in fields(cls):
         field_classification = _classify_interface_field_flavor(
             template_type_hints, template_field)
@@ -2696,35 +517,8 @@ def make_template_definition[T: type](
 
     # Note: this needs to be the absolute last thing, because we need to fully
     # satisfy the intersectable interface before we can call it.
-    _resolve_forward_references(cls)
+    resolve_forward_references(cls)
     return cls
-
-
-def _resolve_forward_references(pending_template_cls: TemplateClass):
-    """The very last thing to do before we return the class after
-    template decoration is to resolve all forward references inside the
-    class. To do that, we first need to construct the corresponding
-    ForwardRefLookupKey and check for it in the pending forward refs
-    lookup.
-
-    If we find one, we then need to update the values there, while
-    checking for and correctly handling recursion.
-    """
-    lookup_key = ForwardRefLookupKey(
-        module=pending_template_cls.__module__,
-        name=pending_template_cls.__name__,
-        scope_id=_extract_frame_scope_id())
-
-    forward_ref_registry = _PENDING_FORWARD_REFS.get()
-    dependent_template_classes = forward_ref_registry.get(lookup_key)
-    if dependent_template_classes is not None:
-        for dependent_template_cls in dependent_template_classes:
-            dependent_xable = cast(
-                TemplateIntersectable, dependent_template_cls)
-            dependent_xable._templatey_signature.resolve_forward_ref(
-                lookup_key, pending_template_cls)
-
-        del forward_ref_registry[lookup_key]
 
 
 @dataclass(frozen=True, slots=True)
